Phase_1/requirements.txt
streamlit
plotly
Pillow
numpy
